Core/Entities: Move GridObject to its own header (#26358)
* GridObject

* Add include to more files
