+++ src/server/game/Entities/GameObject/GameObject.h
@@ -19,6 +19,7 @@
 #define TRINITYCORE_GAMEOBJECT_H
 
 #include "Object.h"
+#include "GridObject.h"
 #include "DatabaseEnvFwd.h"
 #include "GameObjectData.h"
 #include "Loot.h"
