+++ src/server/game/Entities/Object/GridObject.h
@@ -1,0 +1,20 @@
+#ifndef _GRIDOBJECT_H
+#define _GRIDOBJECT_H
+
+#include "GridReference.h"
+#include "GridRefManager.h"
+
+template<class T>
+class GridObject
+{
+    public:
+        virtual ~GridObject() { }
+
+        bool IsInGrid() const { return _gridRef.isValid(); }
+        void AddToGrid(GridRefManager<T>& m) { ASSERT(!IsInGrid()); _gridRef.link(&m, (T*)this); }
+        void RemoveFromGrid() { ASSERT(IsInGrid()); _gridRef.unlink(); }
+    private:
+        GridReference<T> _gridRef;
+};
+
+#endif
